+++ src/cljs/athens/db.cljs
@@ -59,7 +59,6 @@
 (def default-athens-persist
   {:persist/version 1
    :theme/dark      false
-   :window/size     [800 600]
    :graph-conf      default-graph-conf
    :settings        default-settings})
 
@@ -84,7 +83,6 @@
         (update-when-found [:settings :backup-time] "debounce-save-time" js/Number)
         (update-when-found [:settings :monitoring] "monitoring" str->boolean)
         (update-when-found [:theme/dark] "theme/dark" str->boolean)
-        (update-when-found [:window/size] "ws/window-size" #(map js/parseInt (string/split % ",")))
         (update-when-found [:graph-conf] "graph-conf" edn/read-string))))
 
 
@@ -127,7 +125,8 @@
                :mouse-down          false
                :daily-notes/items   []
                :selection           {:items []}
-               :zoom-level          1
+               :theme/dark          false
+               :zoom-level          0
                :fs/watcher          nil
                :presence            {}
                :connection-status   :disconnected})
@@ -220,6 +219,7 @@
 
 
 (defonce dsdb (d/create-conn schema))
+(defonce dsdb-snapshot (atom nil))
 
 
 ;; todo: turn into an effect
@@ -346,7 +346,7 @@
 
 (defn get-parents-recursively
   [id]
-  (->> @(pull dsdb '[:db/id :node/title :block/uid :block/string {:block/_children ...}] id)
+  (->> @(pull dsdb '[:db/id :node/title :block/uid :block/string :edit/time {:block/_children ...}] id)
        shape-parent-query))
 
 
@@ -626,10 +626,10 @@
 (defn group-by-parent
   [blocks]
   (group-by (fn [x]
-              (-> x
-                  :block/parents
-                  first
-                  :node/title))
+              (let [parent (-> x
+                               :block/parents
+                               first)]
+                [(:node/title parent) (:edit/time parent 0)]))
             blocks))
 
 
@@ -646,7 +646,8 @@
        (mapv :db/id)
        merge-parents-and-block
        group-by-parent
-       (sort-by :db/id)
+       (sort-by #(-> % first second))
+       (map #(vector (ffirst %) (second %)))
        vec
        rseq))
 
@@ -658,6 +659,8 @@
        (mapv :db/id)
        merge-parents-and-block
        group-by-parent
+       (sort-by #(-> % first second))
+       (map #(vector (ffirst %) (second %)))
        vec))