+++ src/cljs/athens/views/pages/node_page.cljs
@@ -1,5 +1,6 @@
 (ns athens.views.pages.node-page
   (:require
+    ["/components/Block/components/Anchor" :refer [Anchor]]
     ["/components/Button/Button" :refer [Button]]
     ["/components/Dialog/Dialog" :refer [Dialog]]
     ["@material-ui/core/Popover" :as Popover]
@@ -18,7 +19,6 @@
     [athens.router :refer [navigate-uid navigate]]
     [athens.style :refer [color DEPTH-SHADOWS]]
     [athens.util :refer [escape-str is-daily-note get-caret-position recursively-modify-block-for-embed]]
-    [athens.views.blocks.bullet :as bullet]
     [athens.views.blocks.core :as blocks]
     [athens.views.blocks.textarea-keydown :as textarea-keydown]
     [athens.views.breadcrumbs :refer [breadcrumbs-list breadcrumb]]
@@ -293,7 +293,7 @@
   [parent-uid]
   [:div {:class "block-container"}
    [:div {:style {:display "flex"}}
-    [:span (use-style bullet/bullet-style)]
+    [:> Anchor]
     [:span {:on-click #(handle-new-first-child-block-click parent-uid)} "Click here to add content..."]]])