+++ src/js/components/ProfileSettingsDialog/ProfileSettingsDialog.stories.tsx
@@ -1,8 +1,9 @@
 import React from 'react';
 
+import { useOverlayTriggerState } from "@react-stately/overlays";
+
 import { BADGE, Storybook } from '@/utils/storybook';
 import { ProfileSettingsDialog } from './ProfileSettingsDialog';
-
 
 import { Avatar } from '../Avatar';
 import { Button } from '@/Button';
@@ -20,12 +21,17 @@ export default {
 const testPerson = { personId: '123', username: 'John Doe', color: '#0071ed' };
 
 const Template = (args, context) => {
-  const [isDialogOpen, setIsDialogOpen] = React.useState(args.isOpen);
+  let profileSettingsState = useOverlayTriggerState({
+    defaultOpen: args.defaultOpen,
+    onOpenChange: (open) => {
+      console.log('ProfileSettingsDialog open state changed to: ', open);
+    },
+  });
   const [person, setPerson] = React.useState(testPerson);
 
   const handleUpdatePerson = React.useCallback((person) => {
     setPerson(person);
-    setIsDialogOpen(false);
+    profileSettingsState.close();
   }, []);
 
   return (
@@ -40,24 +46,22 @@ const Template = (args, context) => {
         <Button
           shape="round"
           variant="gray"
-          onClick={() => setIsDialogOpen(true)}>
+          onClick={profileSettingsState.open}>
           Edit
         </Button>
       </div>
       <Storybook.Wrapper>
         <ProfileSettingsDialog
           person={testPerson}
-          isOpen={isDialogOpen}
-          handleUpdatePerson={handleUpdatePerson}
-          handleClose={() => setIsDialogOpen(false)}
+          isOpen={profileSettingsState.isOpen}
+          onClose={profileSettingsState.close}
+          onUpdatePerson={handleUpdatePerson}
         />
       </Storybook.Wrapper>
     </>)
 };
 
 export const Default = Template.bind({});
-
-export const Shown = Template.bind({});
-Shown.args = {
-  isOpen: true
+Default.args = {
+  defaultOpen: true
 }